improve the way messages are displayed
When mailpie-add is run from cron and the output goes to e-mail, the
result is a horrible mess.  this is mostly because progress messages
(useless when you're reading a transcript of the completed program)
clutter the output and obscure real messages.

there are conceptually two kinds of messages: log messages which should
always be displayed, and progress messages which should be displayed
only on a terminal and only until a new log message comes along or the
program ends.

This doesn't fix the biggest offender, which is the output of swish-e
itself.  But it's a step in the right direction.